fix: limpiar lo que denuncia el linter, sin cambiar comportamiento
132 correcciones de ruff en codigo y pruebas. Todas son transformaciones
mecanicas: f-strings en vez de .format(), orden de imports, imports muertos,
elif tras return convertido en if plano, dict() en literales.

Tres cosas que NO son mecanicas y llevan su razon escrita al lado:

- Los seis except Exception a ciegas se marcan con # noqa: BLE001 y su motivo
  en la misma linea, en vez de silenciar la regla desde la configuracion. Los
  seis son deliberados y ya estaban documentados: dos rate limits degradados
  (el del formulario cierra, el del webhook abre a proposito), el health check
  que debe atrapar todo para devolver 503, y tres del patron outbox donde un
  registro malo no puede detener el lote. Ponerlos en ruff.toml los habria
  vuelto invisibles.

- La condicion BAJA de la regla de hinchazon conserva sus dos `if` anidados
  con # noqa: SIM102. No se funden porque el primero pregunta si hay dato y el
  segundo si empeoro, que en clinica son preguntas distintas: "no se" no es
  "no empeoro". Es el mismo razonamiento de D8.

- En settings_production.py el import-estrella vuelve a la primera posicion
  con # isort: skip. El ordenador de imports lo habia movido detras de los
  demas; funcionalmente no rompia nada, pero ese archivo gobierna la seguridad
  de produccion y su orden de carga no se reordena solo.

Revisado a mano el diff de alert_engine.py, bot.py, views.py y los settings
antes de darlo por bueno. Suite: 344 tests OK en 132 s con Django 6.0.8.

+++ Registro_Post_Quirurgico/signos_sintomas/templatetags/panel_admin.py
@@ -102,7 +102,7 @@ def panel_triage(context):
         'respondidos': respondidos,
         'total':       total_checkins,
         'activos':     activos.count(),
-        'pct':         int(round(100 * respondidos / total_checkins)) if total_checkins else 0,
+        'pct':         round(100 * respondidos / total_checkins) if total_checkins else 0,
     }
 
     # Triage acumulado: toda alerta sin resolver (menos SILENCIO), aunque se
